instances salmon and salmon_fillets were NOT merged together
Query: in the Edible Fish class, should the instances 'Salmon fillets' and 'Salmon' be merged together as semantic and ontologic duplicates? yes or no answer only.
Answer:
	 No.

The term "salmon" can refer to various parts of the fish, such as fillets, steaks, or whole fish. In this case, "Salmon fillets" and "Salmon" are not exact synonyms, but rather different forms of the same fish. Therefore, they should not be merged as semantic and ontological duplicates.
meaning that instances salmon and salmon_fillets are different.
Query cache in files:
- /kgfiller/data/cache-3454a5495c5514e3d411176577b84b1423897fdf1786c7de317010d5f097e79d.yml