instances honey and buckwheat_honey were NOT merged together
Query: in the Honey class, should the instances 'Buckwheat honey' and 'Honey' be merged together as semantic and ontologic duplicates? yes or no answer only.
Answer:
	 No.

Buckwheat honey is a type of honey that has a distinct flavor profile and nutritional content compared to other types of honey. It is made from the nectar of buckwheat flowers and has been found to have higher levels of antioxidants and phenolic compounds than other types of honey. Therefore, it would not be appropriate to merge these two instances as they represent different products with unique characteristics.
meaning that instances honey and buckwheat_honey are different.
Query cache in files:
- /kgfiller/data/cache-f15ea02b8df0f6162e7106bb71947e005f6969768c76051c0507a9b20fcc4c5d.yml