instances arborio_rice and rice_wine were NOT merged together
Query: in the Rice class, should the instances 'Arborio rice' and 'Rice wine' be merged together as semantic and ontologic duplicates? yes or no answer only.
Answer:
	 No.

The term "rice" in "Arborio rice" refers to the type of grain used in cooking, whereas the term "rice" in "rice wine" refers to a fermented alcoholic beverage made from rice. These are two distinct concepts and should not be merged as semantic and ontological duplicates.
meaning that instances arborio_rice and rice_wine are different.
Query cache in files:
- /kgfiller/data/cache-7d21e2f9e3723ba297644a84c97a60e81733a23eefb8cab926461e20bcd16e9e.yml